break up conceptset test
